Update tests for create and copy entry APIs
