+++ tests/integration/Entries/CopyEntryTest.cs
@@ -1,4 +1,5 @@
-﻿using Microsoft.VisualStudio.TestTools.UnitTesting;
+﻿using Laserfiche.Api.Client;
+using Microsoft.VisualStudio.TestTools.UnitTesting;
 using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
@@ -66,6 +67,85 @@ namespace Laserfiche.Repository.Api.Client.IntegrationTest.Entries
             var taskCollectionResponse = await client.TasksClient.ListTasksAsync(RepositoryId).ConfigureAwait(false);
             AssertCollectionResponse(taskCollectionResponse);
             Assert.IsNotNull(taskCollectionResponse.Value.First(t => t.Id == opToken && t.Status == TaskStatus.Completed));
+        }
+
+        [TestMethod]
+        public async Task CreateCopyEntry_CopyShortcut()
+        {
+            // Create new entry
+            string newEntryName = "RepositoryApiClientIntegrationTest .Net CreateFolder";
+            int parentEntryId = 1;
+            var createEntryRequest = new CreateEntryRequest()
+            {
+                EntryType = CreateEntryRequestEntryType.Folder,
+                Name = newEntryName,
+                AutoRename = true
+            };
+            var targetEntry = await client.EntriesClient.CreateEntryAsync(RepositoryId, parentEntryId, createEntryRequest).ConfigureAwait(false);
+            Assert.IsNotNull(targetEntry);
+            createdEntries.Add(targetEntry);
+            Assert.AreEqual(parentEntryId, targetEntry.ParentId);
+            Assert.AreEqual(EntryType.Folder, targetEntry.EntryType);
+
+            // Create shortcut to the new entry
+            newEntryName = "RepositoryApiClientIntegrationTest .Net CreateShortcut";
+            createEntryRequest = new CreateEntryRequest()
+            {
+                EntryType = CreateEntryRequestEntryType.Shortcut,
+                Name = newEntryName,
+                TargetId = targetEntry.Id,
+                AutoRename = true
+            };
+            var shortcut = await client.EntriesClient.CreateEntryAsync(RepositoryId, parentEntryId, createEntryRequest).ConfigureAwait(false);
+
+            Assert.IsNotNull(shortcut);
+            createdEntries.Add(shortcut);
+            Assert.AreEqual(parentEntryId, shortcut.ParentId);
+            Assert.AreEqual(EntryType.Shortcut, shortcut.EntryType);
+
+            // Copy entry
+            var copyEntryRequest = new CopyEntryRequest()
+            {
+                Name = "RepositoryApiClientIntegrationTest .Net CopiedEntry",
+                SourceId = shortcut.Id,
+                AutoRename = true
+            };
+            var newEntry = await client.EntriesClient.CopyEntryAsync(RepositoryId, parentEntryId, copyEntryRequest).ConfigureAwait(false);
+
+            createdEntries.Add(newEntry);
+            Assert.IsTrue(newEntry.Name.StartsWith(createEntryRequest.Name));
+            Assert.AreEqual(parentEntryId, newEntry.ParentId);
+            Assert.AreEqual(shortcut.EntryType, newEntry.EntryType);
+        }
+
+
+        [ExpectedException(typeof(ApiException))]
+        [TestMethod]
+        public async Task CreateCopyEntry_CopyFolder()
+        {
+            // Create new entry
+            string newEntryName = "RepositoryApiClientIntegrationTest .Net CreateFolder";
+            int parentEntryId = 1;
+            var createEntryRequest = new CreateEntryRequest()
+            {
+                EntryType = CreateEntryRequestEntryType.Folder,
+                Name = newEntryName,
+                AutoRename = true
+            };
+            var targetEntry = await client.EntriesClient.CreateEntryAsync(RepositoryId, parentEntryId, createEntryRequest).ConfigureAwait(false);
+            Assert.IsNotNull(targetEntry);
+            createdEntries.Add(targetEntry);
+            Assert.AreEqual(parentEntryId, targetEntry.ParentId);
+            Assert.AreEqual(EntryType.Folder, targetEntry.EntryType);
+
+            // Copy entry
+            var copyEntryRequest = new CopyEntryRequest()
+            {
+                Name = "RepositoryApiClientIntegrationTest .Net CopiedEntry",
+                SourceId = targetEntry.Id,
+                AutoRename = true
+            };
+            _ = await client.EntriesClient.CopyEntryAsync(RepositoryId, parentEntryId, copyEntryRequest).ConfigureAwait(false);
         }
     }
 }
